+++ src/cmd/groupby.rs
@@ -53,7 +53,7 @@ You can group on multiple columns (read `xan select -h` for more information abo
 
     $ xan groupby name,surname 'sum(count)' file.csv
 
----
+# Aggregating along columns
 
 This command is also able to aggregate along columns that you can select using
 the --along-cols <cols> flag. In which case, the aggregation functions will accept
@@ -77,6 +77,31 @@ user,count1,count2
 marcy,10,13
 john,4,7
 
+# Aggregating along matrix
+
+This command can also aggregate over all values of a selection of columns, thus
+representing a 2-dimensional matrix, using the -M/--along-matrix flag. In which
+case aggregation functions will accept the anonymous `_` placeholder value representing
+the currently processed column's value.
+
+For instance, given the following file:
+
+user,count1,count2
+marcy,4,5
+john,0,1
+marcy,6,8
+john,4,6
+
+Using the following command:
+
+    $ xan groupby user --along-matrix count1,count2 'sum(_) as total' file.csv
+
+Will produce the following result:
+
+user,total
+marcy,23
+john,11
+
 ---
 
 For a quick review of the capabilities of the expression language,
@@ -96,20 +121,22 @@ Usage:
     xan groupby --help
 
 groupby options:
-    --keep <cols>            Keep this selection of columns, in addition to
-                             the ones representing groups, in the output. Only
-                             values from the first seen row per group will be kept.
-    -C, --along-cols <cols>  Perform a single aggregation over all of selected columns
-                             and create a column per group with the result in the output.
-    -S, --sorted             Use this flag to indicate that the file is already sorted on the
-                             group columns, in which case the command will be able to considerably
-                             optimize memory usage.
-    -p, --parallel           Whether to use parallelization to speed up computation.
-                             Will automatically select a suitable number of threads to use
-                             based on your number of cores. Use -t, --threads if you want to
-                             indicate the number of threads yourself.
-    -t, --threads <threads>  Parellize computations using this many threads. Use -p, --parallel
-                             if you want the number of threads to be automatically chosen instead.
+    --keep <cols>              Keep this selection of columns, in addition to
+                               the ones representing groups, in the output. Only
+                               values from the first seen row per group will be kept.
+    -C, --along-cols <cols>    Perform a single aggregation over all of selected columns
+                               and create a column per group with the result in the output.
+    -M, --along-matrix <cols>  Aggregate all values found in the given selection
+                               of columns.
+    -S, --sorted               Use this flag to indicate that the file is already sorted on the
+                               group columns, in which case the command will be able to considerably
+                               optimize memory usage.
+    -p, --parallel             Whether to use parallelization to speed up computation.
+                               Will automatically select a suitable number of threads to use
+                               based on your number of cores. Use -t, --threads if you want to
+                               indicate the number of threads yourself.
+    -t, --threads <threads>    Parellize computations using this many threads. Use -p, --parallel
+                               if you want the number of threads to be automatically chosen instead.
 
 Common options:
     -h, --help               Display this message
@@ -130,6 +157,7 @@ struct Args {
     flag_delimiter: Option<Delimiter>,
     flag_keep: Option<SelectColumns>,
     flag_along_cols: Option<SelectColumns>,
+    flag_along_matrix: Option<SelectColumns>,
     flag_sorted: bool,
     flag_parallel: bool,
     flag_threads: Option<NonZeroUsize>,
@@ -141,6 +169,10 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
     if args.flag_parallel || args.flag_threads.is_some() {
         if args.flag_along_cols.is_some() {
             Err("-p/--parallel or -t/--threads cannot be used with --along-cols!")?;
+        }
+
+        if args.flag_along_matrix.is_some() {
+            Err("-p/--parallel or -t/--threads cannot be used with --along-matrix!")?;
         }
 
         if args.flag_sorted {
@@ -203,6 +235,46 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
             let group = sel.collect(&record);
 
             program.run_with_cells(group, index, &record, pivot_sel.select(&record))?;
+
+            index += 1;
+        }
+
+        for result in program.into_byte_records(false) {
+            let (group, group_record) = result?;
+
+            write_group(&mut wtr, &group, &group_record)?;
+        }
+
+        return Ok(wtr.flush()?);
+    }
+
+    // --along-matrix
+    if let Some(selection) = args.flag_along_matrix.take() {
+        if args.flag_sorted || args.flag_keep.is_some() {
+            Err("--along-matrix does not work with -S/--sorted nor --keep!")?;
+        }
+
+        let mut matrix_sel = selection.selection(headers, !args.flag_no_headers)?;
+        matrix_sel.sort_and_dedup();
+
+        let mut program =
+            GroupAggregationProgram::<Vec<Vec<u8>>>::parse(&args.arg_expression, headers)?;
+
+        if !args.flag_no_headers {
+            write_group(
+                &mut wtr,
+                &sel.collect(headers),
+                &program.headers().collect(),
+            )?;
+        }
+
+        let mut record = simd_csv::ByteRecord::new();
+        let mut index: usize = 0;
+
+        while rdr.read_byte_record(&mut record)? {
+            let group = sel.collect(&record);
+
+            program.run_with_cells(group, index, &record, matrix_sel.select(&record))?;
 
             index += 1;
         }